pyproject.toml
[tool.poetry]
name = "jwt_course"
version = "0.1.0"
description = ""
authors = ["Matheus Xavier <mxavier+git@rootcert.net>"]

[tool.poetry.dependencies]
python = "^3.7"
PyJWT = "^2.1.0"
dynaconf = "^3.1.4"
PyYAML = "^5.4.1"
fastapi = "^0.65.2"
pymongo = "^3.11.4"
uvicorn = "^0.14.0"
passlib = "^1.7.4"
argon2-cffi = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
